ml_project/requirements.txt
antlr4-python3-runtime==4.8
arrow==1.1.0
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
binaryornot==0.4.4
branca==0.4.1
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.2
coverage==5.5
cycler==0.10.0
dash==1.18.1
dash-bootstrap-components==0.11.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-leaflet==0.1.12
dash-renderer==1.8.3
dash-table==4.11.1
decorator==4.4.2
flake8==3.9.2
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
geobuf==1.1.1
hs-test-python==2.0.1
hydra-core==1.0.6
idna==2.10
importlib-resources==5.1.3
iniconfig==1.1.1
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jinja2-time==0.2.0
joblib==1.0.0
kiwisolver==1.3.1
lightgbm==3.1.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.1
omegaconf==2.0.6
packaging==20.4
pandas==1.2.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.14.1
pluggy==0.13.1
poyo==0.5.0
prompt-toolkit==3.0.8
protobuf==3.14.0
py==1.9.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.7.2
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.12.0
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.0
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
SQLAlchemy==1.3.16
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
